Add `equals.js`
