+++ lib/dropRepeats/index.js
@@ -1,4 +1,4 @@
-import equals from '../equals'
+import equals from '../equals/equals'
 
 export default function dropRepeats(list) {
   var result = []
