+++ lib/whereEq/index.js
@@ -1,5 +1,5 @@
 import _curry2 from '../_internal/_curry2'
-import equals from '../equals'
+import equals from '../equals/equals'
 
 export default _curry2(function whereEq(spec, obj) {
   for (var key in spec) {
